CRUD e viste per la creazione dell'ordine da parte dell'utente verificato

+++ database/migrations/2019_10_15_091108_create_tags_table.php
@@ -26,10 +26,10 @@ class CreateTagsTable extends Migration
         Schema::create('order_tag', function (Blueprint $table) {
             $table->bigIncrements('id');
             
-            $table->unsignedInteger('order_id');
+            $table->unsignedBigInteger('order_id');
             $table->foreign('order_id')->references('id')->on('orders');
 
-            $table->unsignedInteger('tag_id');
+            $table->unsignedBigInteger('tag_id');
             $table->foreign('tag_id')->references('id')->on('tags');
 
             $table->unique(['tag_id', 'order_id']);
